pyproject.toml
[project]
name = "pumpfun-token-creator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anchorpy[cli]>=0.21.0",
    "base58>=2.1.1",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "solana>=0.36.6",
    "solders>=0.26.0",
]
